Manage movement conflict
The case in which the two characters move on the same cell is considered; the lighter of the two is pushed back to the cell on the south; when the two charaters have the same weight, the conflict is resolved in favour of Player1. The logic of how to resolve the conflict needs to be improved.